requirements.txt
torch
torchvision
numpy
opencv-python
scipy
scikit-image
scikit-learn
kornia
einops
numba
Pillow
joblib
h5py
matplotlib
tqdm
transformers
huggingface-hub>0.25
accelerate
timm
PySide6
pyvista
pyvistaqt
open3d
prettytable
tabulate
